Updated the parallel nested sampling
We have removed the messy 'incubator' approach in favour of a larger
'stack' instead of live_data. gestating, daughter and live points are
all stored in the same stack, and this also allows granddaughter points
to be generated.

This has the advantage that we are always trying to generate from the
lowest known contour _including_ that of the points which are
'incubating'. This results in fewer waiting nodes when nprocs~O(nlive).

This apparently works now with nprocs<nlive, but more work is needed to
check this