New attempt at adaptive
I believe the bias occurs because the basis one draws will not align with the
contour perfectly. This causes a concentration of points in the center
when many such directions are drawn across multiple iterations
so that even the spherical case has a bias. However, one does on average
explore the parameter space fully, giving good posteriors and
evidence estimates, but the bias occurs because one is forced inwards

Attempted to correct this with a 'dominant direction' orthogonalisation
procedure, but this doesn't work. does not model the contour well enough
any more once orthogonalised.