Update posterior files
Much less memory is now required, as it only stores the posterior files
that haven't been written to disk

The logic behind storing them on disk is that we're going to be writing
them to disk anyway, so we might as well not bother with the massive
memory requirement

It also allows for a much easier 'trimming' of burn-in, and means that
the posterior points are placed in the order that they are produced. If
we wanted this to occur within memory, this would require a lot of
shuffling on some level, which is expensive with the size of the
posterior files produced.

In addition, the thinning of the chains is now much more consistent, and
should produce much better parameter estimations.