Mid-way through updates
Still at the beginning of nested sampling, but am nearly done updating
the evidence code.

We are bundling all of the relevant information into a single variable
'run_time_info :: RTI'

I have written code to read and resume from this format. Now need to
write code to initialise and use this format.